move getPreview to provider
