+++ src/stelnet/board/query/provider/PeopleProvider.java
@@ -5,11 +5,15 @@ import com.fs.starfarer.api.characters.PersonAPI;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
+import stelnet.board.query.QueryL10n;
 import stelnet.board.query.ResultSet;
 import stelnet.board.query.grouping.GroupingStrategy;
-import stelnet.board.query.view.add.QueryFactory;
 import stelnet.filter.Filter;
 import stelnet.util.CollectionUtils;
+import stelnet.util.L10n;
+import uilib.RenderableComponent;
+import uilib.ShowPeople;
+import uilib.property.Size;
 
 public class PeopleProvider extends QueryProvider {
 
@@ -19,10 +23,6 @@ public class PeopleProvider extends QueryProvider {
         people = null;
     }
 
-    public PeopleProvider(QueryFactory factory) {
-        super(factory);
-    }
-
     @Override
     public List<PersonAPI> getMatching(Set<Filter> filters) {
         List<MarketAPI> markets = MarketProvider.getMarkets(true);
@@ -30,6 +30,11 @@ public class PeopleProvider extends QueryProvider {
         List<PersonAPI> peopleCopy = new LinkedList<>(people);
         CollectionUtils.reduce(peopleCopy, filters);
         return peopleCopy;
+    }
+
+    @Override
+    public RenderableComponent getPreview(Set<Filter> filters, Size size) {
+        return new ShowPeople(getMatching(filters), L10n.get(QueryL10n.NO_MATCHING_PEOPLE), size);
     }
 
     @Override
