+++ src/stelnet/board/query/provider/QueryProvider.java
@@ -7,7 +7,6 @@ import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import stelnet.board.query.ResultSet;
 import stelnet.board.query.grouping.GroupingStrategy;
-import stelnet.board.query.view.add.QueryFactory;
 import stelnet.filter.Filter;
 import stelnet.util.CollectionUtils;
 import uilib.RenderableComponent;
@@ -16,9 +15,9 @@ import uilib.property.Size;
 @RequiredArgsConstructor
 public abstract class QueryProvider {
 
-    private final QueryFactory factory;
-
     public abstract List<?> getMatching(Set<Filter> filters);
+
+    public abstract RenderableComponent getPreview(Set<Filter> filters, Size size);
 
     public List<ResultSet> getResults(Set<Filter> filters, GroupingStrategy groupingStrategy) {
         List<MarketAPI> markets = MarketProvider.getMarkets(true);
@@ -26,10 +25,6 @@ public abstract class QueryProvider {
         List<ResultSet> resultSets = new LinkedList<>();
         processMarkets(resultSets, markets, filters, groupingStrategy);
         return resultSets;
-    }
-
-    public RenderableComponent getPreview(Set<Filter> filters, Size size) {
-        return factory.getPreview(filters, size);
     }
 
     protected abstract void processMarkets(
