+++ src/stelnet/board/query/view/add/ItemQueryFactory.java
@@ -1,6 +1,5 @@
 package stelnet.board.query.view.add;
 
-import com.fs.starfarer.api.campaign.CargoAPI;
 import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -9,20 +8,19 @@ import lombok.Getter;
 import stelnet.CommonL10n;
 import stelnet.board.query.QueryL10n;
 import stelnet.board.query.provider.ItemProvider;
+import stelnet.board.query.view.dialog.ItemPickerDialog;
+import stelnet.board.query.view.dialog.PickerDialog;
 import stelnet.filter.CargoStackNotKnownModspec;
 import stelnet.filter.Filter;
 import stelnet.util.CargoUtils;
 import stelnet.util.L10n;
 import uilib.Button;
 import uilib.Renderable;
-import uilib.RenderableComponent;
-import uilib.ShowCargo;
-import uilib.property.Size;
 
 public class ItemQueryFactory extends QueryFactory {
 
     @Getter
-    private transient ItemProvider provider = new ItemProvider(this);
+    private transient ItemProvider provider = new ItemProvider();
 
     private final FilteringButton[] itemTypes = ItemButtonUtils.createItemTypes();
     private final FilteringButton[] designTypes = ItemButtonUtils.createManufacturers(provider);
@@ -32,7 +30,7 @@ public class ItemQueryFactory extends QueryFactory {
     private final FilteringButton[] wingRoles = ItemButtonUtils.createWingRole();
 
     public Object readResolve() {
-        provider = new ItemProvider(this);
+        provider = new ItemProvider();
         return this;
     }
 
@@ -50,21 +48,12 @@ public class ItemQueryFactory extends QueryFactory {
     }
 
     @Override
-    public RenderableComponent getPreview(Set<Filter> filters, Size size) {
-        return new ShowCargo(
-            getCargo(filters),
-            L10n.get(QueryL10n.MATCHING_ITEMS),
-            L10n.get(QueryL10n.NO_MATCHING_ITEMS),
-            size
-        );
-    }
-
-    @Override
     protected Button[] getFinalComponents() {
         Set<Filter> filters = getFilters(false);
+        PickerDialog picker = new ItemPickerDialog(CargoUtils.makeCargoFromStacks(provider.getMatching(filters)), this);
         return new Button[] {
             new FindMatchingButton(this, L10n.get(CommonL10n.ITEMS)),
-            new FindSelectedButton(this, getCargo(filters)),
+            new FindSelectedButton(picker),
         };
     }
 
@@ -88,10 +77,6 @@ public class ItemQueryFactory extends QueryFactory {
         );
         elements.add(new ButtonGroup(sizeHelper, designTypes, hasWeapons() || hasFighterWings()));
         return elements;
-    }
-
-    private CargoAPI getCargo(Set<Filter> filters) {
-        return CargoUtils.makeCargoFromStacks(provider.getMatching(filters));
     }
 
     private boolean hasWeapons() {
