+++ src/stelnet/board/query/view/add/QueryFactory.java
@@ -17,6 +17,10 @@ public abstract class QueryFactory {
     @Setter
     protected SizeHelper sizeHelper = new SizeHelper();
 
+    public RenderableComponent getPreview(Set<Filter> filters, Size size) {
+        return getProvider().getPreview(filters, size);
+    }
+
     protected void addSelectedOrAll(Set<Filter> filters, FilteringButton buttons[], String type) {
         Set<Filter> selectedFilters = getFilters(buttons, true);
         if (selectedFilters.isEmpty()) {
@@ -47,8 +51,6 @@ public abstract class QueryFactory {
     }
 
     public abstract Set<Filter> getFilters(boolean forResults);
-
-    public abstract RenderableComponent getPreview(Set<Filter> filters, Size size);
 
     public abstract QueryProvider getProvider();
