+++ src/stelnet/board/query/view/add/ShipQueryFactory.java
@@ -1,6 +1,5 @@
 package stelnet.board.query.view.add;
 
-import com.fs.starfarer.api.fleet.FleetMemberAPI;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
@@ -11,18 +10,17 @@ import lombok.Getter;
 import stelnet.CommonL10n;
 import stelnet.board.query.QueryL10n;
 import stelnet.board.query.provider.ShipProvider;
+import stelnet.board.query.view.dialog.PickerDialog;
+import stelnet.board.query.view.dialog.ShipPickerDialog;
 import stelnet.filter.Filter;
 import stelnet.util.L10n;
 import uilib.Button;
 import uilib.Renderable;
-import uilib.RenderableComponent;
-import uilib.ShowShips;
-import uilib.property.Size;
 
 public class ShipQueryFactory extends QueryFactory {
 
     @Getter
-    private transient ShipProvider provider = new ShipProvider(this);
+    private transient ShipProvider provider = new ShipProvider();
 
     private final FilteringButton[] classSizes = ShipButtonUtils.getClassSizes();
     private final FilteringButton[] mountSizes = ShipButtonUtils.getMountSizes();
@@ -33,7 +31,7 @@ public class ShipQueryFactory extends QueryFactory {
     private final FilteringButton[] dMods = ShipButtonUtils.getDMods(provider);
 
     public Object readResolve() {
-        provider = new ShipProvider(this);
+        provider = new ShipProvider();
         return this;
     }
 
@@ -64,21 +62,12 @@ public class ShipQueryFactory extends QueryFactory {
     }
 
     @Override
-    public RenderableComponent getPreview(Set<Filter> filters, Size size) {
-        return new ShowShips(
-            getShips(filters),
-            L10n.get(QueryL10n.MATCHING_SHIPS),
-            L10n.get(QueryL10n.NO_MATCHING_SHIPS),
-            size
-        );
-    }
-
-    @Override
     protected Button[] getFinalComponents() {
         Set<Filter> filters = getFilters(false);
+        PickerDialog picker = new ShipPickerDialog(provider.getMatching(filters), this);
         return new Button[] {
             new FindMatchingButton(this, L10n.get(CommonL10n.SHIPS)),
-            new FindSelectedButton(this, getShips(filters)),
+            new FindSelectedButton(picker),
         };
     }
 
@@ -108,10 +97,6 @@ public class ShipQueryFactory extends QueryFactory {
         addSelectedOrNone(filters, mountBays, L10n.get(QueryL10n.FIGHTER_BAYS), true);
         addSelectedOrNone(filters, designTypes, L10n.get(QueryL10n.MANUFACTURERS), true);
         return filters;
-    }
-
-    private List<FleetMemberAPI> getShips(Set<Filter> filters) {
-        return provider.getMatching(filters);
     }
 
     private void prepareBuiltIns() {
